move integration to cart preview
