+++ modules/checkout/components/CartPreview.vue
@@ -45,6 +45,8 @@
       />
     </div>
     <CouponCode class="highlighted" />
+    <div id="donmo-roundup"></div>
+
     <div class="highlighted">
       <SfCharacteristic
         v-for="characteristic in characteristics"
@@ -59,11 +61,20 @@
 </template>
 <script lang="ts">
 import { SfHeading, SfProperty, SfCharacteristic } from "@storefront-ui/vue";
-import { computed, ref, defineComponent } from "@nuxtjs/composition-api";
+import {
+  computed,
+  ref,
+  defineComponent,
+  onMounted,
+  useContext,
+  watch,
+} from "@nuxtjs/composition-api";
 import cartGetters from "~/modules/checkout/getters/cartGetters";
 import useCart from "~/modules/checkout/composables/useCart";
 import getShippingMethodPrice from "~/helpers/checkout/getShippingMethodPrice";
 import CouponCode from "../../../components/CouponCode.vue";
+
+import { useApi } from "~/composables/useApi";
 
 const CHARACTERISTICS = [
   {
@@ -92,7 +103,7 @@ export default defineComponent({
     CouponCode,
   },
   setup() {
-    const { cart, removeItem, updateItemQty } = useCart();
+    const { cart, removeItem, updateItemQty, load } = useCart();
 
     const listIsHidden = ref(false);
     const products = computed(() => cartGetters.getItems(cart.value));
@@ -107,6 +118,67 @@ export default defineComponent({
       cartGetters.getSelectedShippingMethod(cart.value)
     );
 
+    const { mutate } = useApi();
+    const {
+      app: { i18n },
+    } = useContext();
+
+    onMounted(() => {
+      load().then(() => {
+        const donmo = (window as any).DonmoRoundup({
+          publicKey: process.env.DONMO_PUBLIC_KEY,
+          isBackendBased: true,
+          language: i18n.locale,
+          orderId: cart.value.id,
+          addDonationAction: async ({ donationAmount }) => {
+            const ADD_DONATION_MUTATION = `
+              mutation ADD_DONATION_MUTATION($donationAmount: Float!, $cartId: String!) {
+                addDonationToQuote(donationAmount: $donationAmount, cartId: $cartId){
+                  message
+                }
+              }
+            `;
+
+            await mutate(ADD_DONATION_MUTATION, {
+              donationAmount,
+              cartId: cart.value.id,
+            });
+
+            // refresh cart
+            load();
+          },
+          removeDonationAction: async () => {
+            const REMOVE_DONATION_MUTATION = `
+              mutation REMOVE_DONATION_MUTATION($cartId: String!) {
+                removeDonationFromQuote(cartId: $cartId) {
+                  message
+                }
+              }
+            `;
+
+            await mutate(REMOVE_DONATION_MUTATION, { cartId: cart.value.id });
+
+            // refresh cart
+            load();
+          },
+          getExistingDonation: () => {
+            return cart.value?.prices?.["donmo_donation"]?.value;
+          },
+          getGrandTotal: () => cartGetters.getTotals(cart.value).total,
+        });
+
+        donmo.build();
+
+        watch(
+          () => cartGetters.getTotals(cart.value).total,
+          () => {
+            load().then(() => {
+              donmo.refresh();
+            });
+          }
+        );
+      });
+    });
     return {
       cart,
       discount,
@@ -124,10 +196,26 @@ export default defineComponent({
       selectedShippingMethod,
     };
   },
+
+  head() {
+    return {
+      title: "Donmo Roundup", // Other meta information
+      script: [
+        {
+          hid: "donmo",
+          src: "https://static.donmo.org/integration.js",
+          defer: true,
+        },
+      ],
+    };
+  },
 });
 </script>
 
 <style lang="scss" scoped>
+#donmo-roundup {
+  margin-bottom: 10px;
+}
 .highlighted {
   box-sizing: border-box;
   width: 100%;
